fix: switched scope for job triggers
It looks like it's not possible to ask questions about a pop/pop_group from inside the triggered_planet_modifier scope. In the debug report, the local scope is just the planet, with no reference to prev/from/root. The current solution here checks the owner's primary species traits, which will often be what you want but not always.